+++ CHANGELOG.md
@@ -1,2 +1,4 @@
-# 2.1.0 Mar 13, 2021
-* Updated Plugin Requirements.
+# 2.1.0 Mar 13, 2023
+* Updated Plugin Requirements
+* Added Gender and Age group taxonomies to product
+* Allowed product content to accept blocks.